Update Gov.uk data to 2010-04-10, fix key name change
Downloaded the latest daily indicators Excel file from Gov.uk, opening in
LibreOffice and saving as a text CSV file.

The DailyUKCases heading had changed to NewUKCases, so the value of this heading
does not work with the convention used in the headline figure card helper
functions.  A quick fix was to add the previous key using the value from the new
key.